+++ packages/vseed/src/pipeline/spec/chart/pipes/init/donut.ts
@@ -13,8 +13,8 @@ export const initDonut: SpecPipe = (spec, context) => {
   result.outerRadius = 0.8
   result.innerRadius = result.outerRadius * 0.8
   result.valueField = foldInfo.measureValue
-  result.categoryField = unfoldInfo.encodingDetail
-  result.seriesField = unfoldInfo.encodingColorId
+  result.categoryField = unfoldInfo.encodingColorId
+
   result.padding = 0
   result.region = [
     {